+++ lib/pages/general/generators.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:realm_roller/custom_widgets/route_builder/route_builder.dart';
 import 'package:realm_roller/pages/locations/locations_generation/location_generation_page.dart';
-import 'package:realm_roller/pages/names/names_generation/names_generation_page.dart';
+import 'package:realm_roller/pages/names/names_generation/npc_names_generation_page.dart';
 import 'package:realm_roller/pages/npcs/npc_generation/npc_generation_page.dart';
+import 'package:realm_roller/pages/settlements/settlements_generation/settlement_generation_page.dart';
+
+import '../names/names_generation/settlement_name_generation.dart';
 
 /// A widget that lists all the available generators
 class GeneratorsPage extends StatelessWidget {
@@ -12,8 +15,10 @@ class GeneratorsPage extends StatelessWidget {
   Widget build(BuildContext context) {
     const generators = {
       "Npc": NpcGenerationPage(),
-      "Names": NamesGenerationPage(),
-      "Locations": LocationGenerationPage()
+      "Names": NpcNamesGenerationPage(),
+      "Locations": LocationGenerationPage(),
+      "Settlements": SettlementGenerationPage(),
+      "Settlement Names": SettlementNamesGenerationPage(),
     };
     final titles = generators.keys.toList();